# api_doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "deaftone",
    "description": "## About \nThis api is inspired by https://auraspec.readthedocs.io/en/latest/api.html\n### Response Format and Errors\n\nThe MIME type for all responses MUST be ``application/vnd.api+json``. Every response is a JSON object. When a request is successful, the document has a top-level key data corresponding to the response’s “primary data.” When it fails, the document has an errors key, which maps to an array of JSON API error objects. Other keys may also be present, as described below.\n\nOn a successfuly query\n``\n{\n    status: \"success\"\n    data: {Response}\n}\n``\nOn a failure\n``\n{\n    status: \"error\",\n    message: String,\n}``\n        ",
    "license": {
      "name": ""
    },
    "version": "0.0.3"
  },
  "paths": {
    "/albums": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::albums"
        ],
        "operationId": "get_albums",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List containing albums",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumsResponseOpenApi"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get albums",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/albums/{album_id}": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::albums"
        ],
        "operationId": "get_album",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "description": "Album Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a album",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseOpenApi"
                }
              }
            }
          },
          "404": {
            "description": "Album not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error occured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/albums/{album_id}/cover": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::albums"
        ],
        "operationId": "get_cover",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "description": "Album Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a album cover"
          },
          "404": {
            "description": "Album not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/artists": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::artists"
        ],
        "operationId": "get_artists",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List containing artists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistsResponseOpenApi"
                }
              }
            }
          },
          "500": {
            "description": "Error occured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{artist_id}": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::artists"
        ],
        "operationId": "get_artist",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "description": "Artist Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponseOpenApi"
                }
              }
            }
          },
          "404": {
            "description": "Failed to get artist ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error occured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{playlist_id}": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::playlist"
        ],
        "operationId": "get_playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "description": "Playlist Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a playlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayListResponseOpenApi"
                }
              }
            }
          },
          "404": {
            "description": "Playlist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error occured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/song/{song_id}": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::songs"
        ],
        "operationId": "get_song",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "description": "Song Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/stream/{song_id}": {
      "get": {
        "tags": [
          "deaftone::services::http::handlers::streams"
        ],
        "operationId": "stream_handler",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "description": "Song Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a song stream",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoxBody"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlbumResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "artist",
          "artist_id",
          "album_description",
          "year",
          "song_count",
          "songs"
        ],
        "properties": {
          "album_description": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "artist_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "song_count": {
            "type": "integer",
            "format": "int32"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity.song.Model"
            }
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AlbumResponseOpenApi": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "$ref": "#/components/schemas/AlbumResponse"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
      "AlbumsResponseOpenApi": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity.album.Model"
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
      "ArtistLinks": {
        "type": "object",
        "properties": {
          "all_music": {
            "type": "string",
            "nullable": true
          },
          "amazon_music": {
            "type": "string",
            "nullable": true
          },
          "apple_music": {
            "type": "string",
            "nullable": true
          },
          "deezer": {
            "type": "string",
            "nullable": true
          },
          "discogs": {
            "type": "string",
            "nullable": true
          },
          "facebook": {
            "type": "string",
            "nullable": true
          },
          "itunes": {
            "type": "string",
            "nullable": true
          },
          "spotify": {
            "type": "string",
            "nullable": true
          },
          "tidal": {
            "type": "string",
            "nullable": true
          },
          "twitter": {
            "type": "string",
            "nullable": true
          },
          "wiki": {
            "type": "string",
            "nullable": true
          },
          "youtube": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ArtistResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "image",
          "biography",
          "links",
          "albums"
        ],
        "properties": {
          "albums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity.album.Model"
            }
          },
          "biography": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "links": {
            "$ref": "#/components/schemas/ArtistLinks"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ArtistResponseOpenApi": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ArtistResponse"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
      "ArtistsResponseOpenApi": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity.artist.Model"
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "PlayListResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "songs"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity.song.Model"
            }
          }
        }
      },
      "PlayListResponseOpenApi": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "$ref": "#/components/schemas/PlayListResponse"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
      "SongResponse": {
        "type": "object",
        "required": [
          "id",
          "path",
          "title",
          "disk",
          "artist",
          "album_name",
          "length",
          "year",
          "album_id",
          "liked"
        ],
        "properties": {
          "album_id": {
            "type": "string"
          },
          "album_name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "disk": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "length": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "liked": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Status": {
        "type": "string",
        "enum": [
          "success",
          "error"
        ]
      },
      "entity.album.Model": {
        "type": "object",
        "required": [
          "id",
          "path",
          "name",
          "year",
          "artist_name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "album_artist": {
            "type": "string",
            "nullable": true
          },
          "album_artist_credit": {
            "type": "string",
            "nullable": true
          },
          "album_artist_sort": {
            "type": "string",
            "nullable": true
          },
          "album_disambig": {
            "type": "string",
            "nullable": true
          },
          "album_status": {
            "type": "string",
            "nullable": true
          },
          "album_type": {
            "type": "string",
            "nullable": true
          },
          "artist_id": {
            "type": "string",
            "nullable": true
          },
          "artist_name": {
            "type": "string"
          },
          "asin": {
            "type": "string",
            "nullable": true
          },
          "catalog_num": {
            "type": "string",
            "nullable": true
          },
          "composer": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "cover": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "discogs_albumid": {
            "type": "string",
            "nullable": true
          },
          "discogs_artistid": {
            "type": "string",
            "nullable": true
          },
          "discogs_labelid": {
            "type": "string",
            "nullable": true
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "example": "Album"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "mb_album_id": {
            "type": "string",
            "nullable": true
          },
          "mb_artist_id": {
            "type": "string",
            "nullable": true
          },
          "mb_releasegroup_id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "release_group_disambig": {
            "type": "string",
            "nullable": true
          },
          "script": {
            "type": "string",
            "nullable": true
          },
          "style": {
            "type": "string",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "entity.artist.Model": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "biography": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "example": "Artist"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "link_all_music": {
            "type": "string",
            "nullable": true
          },
          "link_amazon_music": {
            "type": "string",
            "nullable": true
          },
          "link_apple_music": {
            "type": "string",
            "nullable": true
          },
          "link_deezer": {
            "type": "string",
            "nullable": true
          },
          "link_discogs": {
            "type": "string",
            "nullable": true
          },
          "link_facebook": {
            "type": "string",
            "nullable": true
          },
          "link_itunes": {
            "type": "string",
            "nullable": true
          },
          "link_spotify": {
            "type": "string",
            "nullable": true
          },
          "link_tidal": {
            "type": "string",
            "nullable": true
          },
          "link_twitter": {
            "type": "string",
            "nullable": true
          },
          "link_wiki": {
            "type": "string",
            "nullable": true
          },
          "link_youtube": {
            "type": "string",
            "nullable": true
          },
          "mb_artist_id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}